+++ airlab/transformation/pairwise.py
@@ -34,11 +34,24 @@ class _Transformation(th.nn.Module):
         self._image_size = np.array(image_size)
         self._constant_displacement = None
         self._diffeomorphic = diffeomorphic
+        self._constant_flow = None
+
+        self._compute_flow = None
 
         if self._diffeomorphic:
             self._diffeomorphic_calculater = tu.Diffeomorphic(image_size, dtype=dtype, device=device)
         else:
             self._diffeomorphic_calculater = None
+
+    def get_flow(self):
+
+        if self._constant_flow is None:
+            return self._compute_flow().detach()
+        else:
+            return self._compute_flow().detach() + self._constant_flow
+
+    def set_constant_flow(self, flow):
+        self._constant_flow = flow
 
     def get_displacement_numpy(self):
 
@@ -48,42 +61,53 @@ class _Transformation(th.nn.Module):
             return self().detach().cpu().numpy()
 
     def get_displacement(self):
-
             return self().detach()
 
-    def get_current_displacement(self):
+    # def get_current_displacement(self):
+    #
+    #     if self._dim == 2:
+    #         return th.unsqueeze(self().detach(), 0).cpu().numpy()
+    #     elif self._dim == 3:
+    #         return self().detach().cpu().numpy()
 
-        if self._dim == 2:
-            return th.unsqueeze(self().detach(), 0).cpu().numpy()
-        elif self._dim == 3:
-            return self().detach().cpu().numpy()
+    # def set_constant_displacement(self, displacement):
+    #
+    #     self._constant_displacement = displacement
 
-    def set_constant_displacement(self, displacement):
+    # def get_inverse_transformation(self, displacement):
+    #     if self._diffeomorphic:
+    #         if self._dim == 2:
+    #             inv_displacement = self._diffeomorphic_calculater.calculate(displacement * -1)
+    #         else:
+    #             inv_displacement = self._diffeomorphic_calculater.calculate(displacement * -1)
+    #     else:
+    #         print("error displacement ")
+    #         inv_displacement = None
+    #
+    #     return inv_displacement
 
-        self._constant_displacement = displacement
+    def get_inverse_displacement(self):
 
-    def get_inverse_transformation(self, displacement):
+        flow = self._concatenate_flows(self._compute_flow()).detach()
+
         if self._diffeomorphic:
-            if self._dim == 2:
-                inv_displacement = self._diffeomorphic_calculater.calculate(displacement * -1)
-            else:
-                inv_displacement = self._diffeomorphic_calculater.calculate(displacement * -1)
+                inv_displacement = self._diffeomorphic_calculater.calculate(flow * -1)
         else:
             print("error displacement ")
             inv_displacement = None
 
         return inv_displacement
 
-    def _diffeomorphic_transformation(self, displacement):
+    def _compute_diffeomorphic_displacement(self, flow):
 
-        return self._diffeomorphic_calculater.calculate(displacement)
+        return self._diffeomorphic_calculater.calculate(flow)
 
-    def _return_displacement(self, displacement):
+    def _concatenate_flows(self, flow):
 
-        if self._constant_displacement is None:
-            return displacement
+        if self._constant_flow is None:
+            return flow
         else:
-            return (displacement + self._constant_displacement)
+            return flow + self._constant_flow
 
 
 class RigidTransformation(_Transformation):
@@ -212,7 +236,7 @@ class RigidTransformation(_Transformation):
                                                   self._rotation_matrix), self._trans_matrix_cm_rw)[0:self._dim, :]
         return transformation_matrix
 
-    def _compute_dense_displacement(self, transformation_matrix):
+    def _compute_dense_flow(self, transformation_matrix):
 
         displacement = th.mm(self._grid.view(np.prod(self._image_size).tolist(), self._dim + 1),
                              transformation_matrix.t()).view(*(self._image_size.tolist()), self._dim) \
@@ -227,9 +251,9 @@ class RigidTransformation(_Transformation):
 
         self._compute_transformation()
         transformation_matrix = self._compute_transformation_matrix()
-        displacement = self._compute_dense_displacement(transformation_matrix)
+        flow = self._compute_dense_flow(transformation_matrix)
 
-        return self._return_displacement(displacement)
+        return self._concatenate_flows(flow)
 
 
 class SimilarityTransformation(RigidTransformation):
@@ -284,9 +308,9 @@ class SimilarityTransformation(RigidTransformation):
 
         self._compute_transformation()
         transformation_matrix = self._compute_transformation_matrix()
-        displacement = self._compute_dense_displacement(transformation_matrix)
+        flow = self._compute_dense_flow(transformation_matrix)
 
-        return self._return_displacement(displacement)
+        return self._concatenate_flows(flow)
 
 
 class AffineTransformation(SimilarityTransformation):
@@ -348,10 +372,9 @@ class AffineTransformation(SimilarityTransformation):
 
         self._compute_transformation()
         transformation_matrix = self._compute_transformation_matrix()
-        displacement = self._compute_dense_displacement(transformation_matrix)
+        flow = self._compute_dense_flow(transformation_matrix)
 
-        return self._return_displacement(displacement)
-
+        return self._concatenate_flows(flow)
 
 
 class NonParametricTransformation(_Transformation):
@@ -369,9 +392,9 @@ class NonParametricTransformation(_Transformation):
         self.to(dtype=self._dtype, device=self._device)
 
         if self._dim == 2:
-            self._compute_displacement = self._compute_displacement_2d
+            self._compute_flow = self._compute_flow_2d
         else:
-            self._compute_displacement = self._compute_displacement_3d
+            self._compute_flow = self._compute_flow_3d
 
     def set_start_parameter(self, parameters):
         if self._dim == 2:
@@ -380,17 +403,19 @@ class NonParametricTransformation(_Transformation):
             self.trans_parameters = Parameter(th.tensor(parameters.transpose(0, 1)
                                                         .transpose(0, 2).transpose(0, 3)))
 
-    def _compute_displacement_2d(self):
+    def _compute_flow_2d(self):
         return self.trans_parameters.transpose(0, 2).transpose(0, 1)
 
-    def _compute_displacement_3d(self):
+    def _compute_flow_3d(self):
         return self.trans_parameters.transpose(0, 3).transpose(0, 2).transpose(0, 1)
 
     def forward(self):
-        displacement = self._return_displacement(self._compute_displacement())
+        flow = self._concatenate_flows(self._compute_flow())
 
         if self._diffeomorphic:
-            displacement = self._diffeomorphic_transformation(displacement)
+            displacement = self._compute_diffeomorphic_displacement(flow)
+        else:
+            displacement = flow
 
         return displacement
 
@@ -410,9 +435,10 @@ class _KernelTransformation(_Transformation):
         assert self._dim == 2 or self._dim == 3
 
         if self._dim == 2:
-            self._compute_displacement = self._compute_displacement_2d
+            self._compute_flow = self._compute_flow_2d
         else:
-            self._compute_displacement = self._compute_displacement_3d
+            self._compute_flow = self._compute_flow_3d
+
 
     def get_current_displacement(self):
 
@@ -463,7 +489,7 @@ class _KernelTransformation(_Transformation):
         size = [1, 1] + self._image_size.astype(dtype=int).tolist()
         self._displacement = th.empty(*size, dtype=self._dtype, device=self._device)
 
-    def _compute_displacement_2d(self):
+    def _compute_flow_2d(self):
         displacement_tmp = F.conv_transpose2d(self.trans_parameters, self._kernel,
                                           padding=self._padding, stride=self._stride, groups=2)
 
@@ -472,7 +498,7 @@ class _KernelTransformation(_Transformation):
                        self._stride[0] + self._crop_start[0]:-self._stride[0] - self._crop_end[0],
                        self._stride[1] + self._crop_start[1]:-self._stride[1] - self._crop_end[1]].transpose_(1, 3).transpose(1, 2))
 
-    def _compute_displacement_3d(self):
+    def _compute_flow_3d(self):
 
         # compute dense displacement
         displacement = F.conv_transpose3d(self.trans_parameters, self._kernel,
@@ -486,10 +512,12 @@ class _KernelTransformation(_Transformation):
 
     def forward(self):
 
-        displacement = self._return_displacement(self._compute_displacement())
+        flow = self._concatenate_flows(self._compute_flow())
 
         if self._diffeomorphic:
-            displacement = self._diffeomorphic_transformation(displacement)
+            displacement = self._compute_diffeomorphic_displacement(flow)
+        else:
+            displacement = flow
 
         return displacement